try to add bvhs, moved onto global illumination

+++ src/Example/Example.h
@@ -16,7 +16,7 @@ public:
 	virtual void draw() = 0;
 protected:
 	int index(int x, int y);
-	Ray getRayThroughPixel(int x, int y);
+	Ray getRayThroughPixel(float x, float y);
 	void setPixelColor(int x, int y, const glm::vec3& color);
 
 	//depth test, check if d is closer to the camera than what is currently stored
